# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Restaurant Bucket List API
  version: 1.0.0
paths:
  /restaurants/:
    get:
      summary: Get a list of all restaurants
      responses:
        '200':
          description: A list of restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Restaurant'
    post:
      summary: Add a new restaurant to the list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Restaurant'
      responses:
        '201':
          description: New restaurant added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'

  /restaurants/{id}:
    get:
      summary: Get a specific restaurant by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A single restaurant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
    put:
      summary: Update a restaurant's details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Restaurant'
      responses:
        '200':
          description: Restaurant updated
    delete:
      summary: Delete a restaurant from the list
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Restaurant deleted

components:
  schemas:
    Restaurant:
      type: object
      required:
        - name
        - city
        - cuisine
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
        city:
          type: string
        cuisine:
          type: string